requirements.txt
pydantic
pydantic-settings
httpx
python-dotenv
pandas
google-cloud-storage
pytest
